# apps/docs/openapi/writing-app-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Writing App API",
    "description": "개발 문서",
    "version": "0.0.1"
  },
  "paths": {
...
    "/courses/{courseId}/curriculum-upgrade": {
      "get": {
        "operationId": "getCoursesByCourseIdCurriculumUpgrade",
        "responses": {
          "200": {
            "description": "현재 학습자의 커리큘럼 업그레이드 공지입니다.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "courseId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "status": {
                          "type": "string",
                          "const": "not-available"
                        }
                      },
                      "required": ["courseId", "status"]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "completedCount": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        },
                        "courseId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "fromVersion": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "minLength": 1
                            },
                            "title": {
                              "type": "string",
                              "minLength": 1
                            },
                            "versionNumber": {
                              "type": "integer",
                              "exclusiveMinimum": 0,
                              "maximum": 9007199254740991
                            }
                          },
                          "required": ["id", "title", "versionNumber"]
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        },
                        "migrationId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "status": {
                          "type": "string",
                          "const": "available"
                        },
                        "toVersion": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "minLength": 1
                            },
                            "title": {
                              "type": "string",
                              "minLength": 1
                            },
                            "versionNumber": {
                              "type": "integer",
                              "exclusiveMinimum": 0,
                              "maximum": 9007199254740991
                            },
                            "changelog": {
                              "type": "string",
                              "minLength": 1
                            }
                          },
                          "required": [
                            "id",
                            "title",
                            "versionNumber",
                            "changelog"
                          ]
                        },
                        "totalLessons": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "completedCount",
                        "courseId",
                        "fromVersion",
                        "message",
                        "migrationId",
                        "status",
                        "toVersion",
                        "totalLessons"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "unauthorized"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "404": {
            "description": "코스를 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "course-not-found"
                    },
                    "message": {
                      "type": "string",
                      "const": "코스를 찾을 수 없습니다."
                    },
                    "courseId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message", "courseId"]
                }
              }
            }
          },
          "500": {
            "description": "콘텐츠 시드가 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-content-seed"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "503": {
            "description": "데이터베이스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "database-unavailable"
                    },
                    "message": {
                      "type": "string",
                      "const": "데이터베이스를 사용할 수 없습니다."
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "courseId",
            "required": true
          }
        ]
      },
      "post": {
        "operationId": "postCoursesByCourseIdCurriculumUpgrade",
        "responses": {
          "200": {
            "description": "현재 학습자의 커리큘럼 업그레이드를 적용했습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "completedLessonCount": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "completedLessonIds": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "courseId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "createdAt": {
                      "type": "string",
                      "minLength": 1
                    },
                    "fromVersionId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "id": {
                      "type": "string",
                      "minLength": 1
                    },
                    "migrationId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "preservedLessonIds": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "skippedLessonIds": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "status": {
                      "type": "string",
                      "const": "completed"
                    },
                    "toVersionId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "updatedAt": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "completedLessonCount",
                    "completedLessonIds",
                    "courseId",
                    "createdAt",
                    "fromVersionId",
                    "id",
                    "migrationId",
                    "preservedLessonIds",
                    "skippedLessonIds",
                    "status",
                    "toVersionId",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "업그레이드 요청이 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-request"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "unauthorized"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "404": {
            "description": "코스 또는 커리큘럼 업그레이드를 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "const": "course-not-found"
                        },
                        "message": {
                          "type": "string",
                          "const": "코스를 찾을 수 없습니다."
                        },
                        "courseId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": ["code", "message", "courseId"]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "const": "not-found"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": ["code", "message"]
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "콘텐츠 시드가 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-content-seed"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "503": {
            "description": "데이터베이스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "database-unavailable"
                    },
                    "message": {
                      "type": "string",
                      "const": "데이터베이스를 사용할 수 없습니다."
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "courseId",
            "required": true
          }
        ]
      }
    },
    "/courses/{courseId}/curriculum-upgrade/dismiss": {
      "post": {
        "operationId": "postCoursesByCourseIdCurriculumUpgradeDismiss",
        "responses": {
          "200": {
            "description": "현재 학습자의 커리큘럼 업그레이드 공지를 숨겼습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "courseId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "dismissedAt": {
                      "type": "string",
                      "minLength": 1
                    },
                    "fromVersionId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "string",
                      "const": "dismissed"
                    },
                    "toVersionId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": [
                    "courseId",
                    "dismissedAt",
                    "fromVersionId",
                    "status",
                    "toVersionId"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "로그인이 필요합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "unauthorized"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "404": {
            "description": "코스 또는 커리큘럼 업그레이드를 찾을 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "const": "course-not-found"
                        },
                        "message": {
                          "type": "string",
                          "const": "코스를 찾을 수 없습니다."
                        },
                        "courseId": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": ["code", "message", "courseId"]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "const": "not-found"
                        },
                        "message": {
                          "type": "string",
                          "minLength": 1
                        }
                      },
                      "required": ["code", "message"]
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "콘텐츠 시드가 올바르지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "invalid-content-seed"
                    },
                    "message": {
                      "type": "string",
                      "minLength": 1
                    },
                    "lessonId": {
                      "type": "string",
                      "minLength": 1
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          },
          "503": {
            "description": "데이터베이스를 사용할 수 없습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "database-unavailable"
                    },
                    "message": {
                      "type": "string",
                      "const": "데이터베이스를 사용할 수 없습니다."
                    }
                  },
                  "required": ["code", "message"]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "courseId",
            "required": true
          }
        ]
      }
    },
...
  },
...
}